refactor: Internal type cleanup
