+++ src/definitions/enumType.ts
@@ -1,6 +1,6 @@
 import { assertValidName, GraphQLEnumTypeConfig, GraphQLEnumValueConfig } from 'graphql'
 import { arg, NexusArgDef, NexusAsArgConfig } from './args'
-import { NexusTypes, SourceTypingDef, withNexusSymbol } from './_types'
+import { Maybe, NexusTypes, SourceTypingDef, withNexusSymbol } from './_types'
 
 type TypeScriptEnumLike = {
   [key: number]: string
@@ -12,12 +12,12 @@ export interface EnumMemberInfo {
   /** The internal representation of the enum */
   value?: string | number | object | boolean
   /** The description to annotate the GraphQL SDL */
-  description?: string
+  description?: Maybe<string>
   /**
    * Info about a field deprecation. Formatted as a string and provided with the deprecated directive on
    * field/enum types and as a comment on input fields.
    */
-  deprecation?: string // | DeprecationInfo;
+  deprecation?: Maybe<string> // | DeprecationInfo;
   /**
    * Custom extensions, as supported in graphql-js
    *
@@ -29,7 +29,7 @@ export interface EnumMemberInfo {
 export interface NexusEnumTypeConfig<TypeName extends string> {
   name: TypeName
   /** The description to annotate the GraphQL SDL */
-  description?: string
+  description?: Maybe<string>
   /** Source type information for this type */
   sourceType?: SourceTypingDef
   /** All members of the enum, either as an array of strings/definition objects, as an object, or as a TypeScript enum */
