+++ src/definitions/inputObjectType.ts
@@ -1,7 +1,7 @@
 import { assertValidName, GraphQLInputObjectTypeConfig } from 'graphql'
 import { arg, NexusArgDef, NexusAsArgConfig } from './args'
 import type { InputDefinitionBlock } from './definitionBlocks'
-import { NexusTypes, NonNullConfig, withNexusSymbol } from './_types'
+import { Maybe, NexusTypes, NonNullConfig, withNexusSymbol } from './_types'
 
 export type NexusInputObjectTypeConfig<TypeName extends string> = {
   /** Name of the input object type */
@@ -9,7 +9,7 @@ export type NexusInputObjectTypeConfig<TypeName extends string> = {
   /** Definition block for the input type */
   definition(t: InputDefinitionBlock<TypeName>): void
   /** The description to annotate the GraphQL SDL */
-  description?: string
+  description?: Maybe<string>
   /**
    * Configures the nullability for the type, check the documentation's "Getting Started" section to learn
    * more about GraphQL Nexus's assumptions and configuration on nullability.
